feat: restructuring codes! moving neo4j utils into its right directory.
