+++ discord_analyzer/analysis/neo4j_analysis/analyzer_node_stats.py
@@ -3,7 +3,7 @@ import logging
 from uuid import uuid1
 
 import pandas as pd
-from discord_analyzer.analysis.neo4j_utils.projection_utils import ProjectionUtils
+from discord_analyzer.analysis.neo4j_analysis.utils import ProjectionUtils
 from tc_neo4j_lib.neo4j_ops import Neo4jOps
