+++ discord_analyzer/analysis/neo4j_analysis/centrality.py
@@ -2,8 +2,7 @@ import logging
 from typing import Literal
 
 import pandas as pd
-from discord_analyzer.analysis.neo4j_utils.neo4j_metrics import Neo4JMetrics
-from discord_analyzer.analysis.neo4j_utils.projection_utils import ProjectionUtils
+from discord_analyzer.analysis.neo4j_analysis.utils import Neo4JMetrics, ProjectionUtils
 from tc_neo4j_lib.neo4j_ops import Neo4jOps, Query
